Start of a hubot for deploying.
So far, the monkey (sun wukong, the monkey king) just listens for
commands and responds to hipchat if it hears them.

Pair programmed with kamens:

Auditors: ben

Test Plan:
Ran
   bin/culturecow
manually on toby with this new code, and did
   sun, ping
in both the deploy room and 1s/0s, and got the expected response.

I also did
   sun, deploy branch foo
in the deploy room, and it said:
   Telling hipchat to deploy branch foo

and then
   Sun, deploy bar
and it said
   Telling hipchat to deploy branch foo