Allow user to force a deploy-start even if sun doesn't think it's time for one.
Summary:
Sun could get messed up, and before this there was no way to get him
back on track except by restarting the culture cow.  This should serve
as an easier way; the 'dagnabit' forces sun to start from the top.

We can't use dagnabit on the other commands, because those depend on
state that sun has read from the hipchat logs, and if sun is messed up
it won't have that state.

Test Plan: Fingers crossed.

Reviewers: ben

Reviewed By: ben

Differential Revision: http://phabricator.khanacademy.org/D13179