Add support for an 'emergency rollback' command.
Summary:
This is equivalent to running 'deploy/set_default.py rollback' on the
commandline, and rolls back the live site to a previous, 'safe'
version (and marks the current version as bad inside git).

Test Plan:
Fingers crossed.  Once this is deployed, I'll say
   sun, rollback
in the '1s/0s: deploys' room, and will also change the EMERGENCY
ROLLBACK jenkins job to temporarily add the '--dry-run' flag to
set-default.py, and will then say
   sun, emergency rollback

Reviewers: ben

Reviewed By: ben

Subscribers: pamela

Differential Revision: http://phabricator.khanacademy.org/D14134