Adpated hubot scripts to work with hipchat.
Summary:
Most scripts I just changed what they listen to.  Since we really only
have just one bot (Culture Cow) running, I changed the respond's to
hear's and just made their regular expressions fairly stringent so that
people didn't accidentily trigger a bot when they did not mean to.

When the command tt is run, the bot prints out a list of commands, so I
had to do a check in a bunch of scripts to make sure that the bot
listing out the commands didnt trigger the command itself.

I also changed the msg.send's to use the fancyMessage so that we respond
in the correct room for the message.

There were other more substaintal changes such as with wolfram where the
node module wasn't working but I found another module that did work.

Hipchat didn't work well with lists of things to be printed so with
Zendesk I had to modify to send the list as one message.

For the command list to work I added the correct format for docstrings
that hubot wants.

Test Plan:
in secrets.py added the new secrets, and changed the room to jamestest
in scripts/hipchat.py fancyMessage added "jamestest": "jamestest" in the
rooms dict.
Then ran:
bin/culture-cow

In jamestest hipchat room:
ran tt and then all the other listed commands to see that they work as
expected.

Reviewers: csilvers

Reviewed By: csilvers

Differential Revision: http://phabricator.khanacademy.org/D14264