Get a full list of rooms in hipchat-api.js, and fix the brain-path.
Summary:
I got tired of culture-cow failing to send stuff to a room because I
forgot to update the xmpp -> name mapping.  So now I just hard-code in
all the mappings (even though 90+% of them will never be used), and
add instructions for how to easily update these mappings in the
future.

I also saw the following error in the hubot logs:
   ERROR Error: ENOENT, no such file or directory 'data/brain-dump.json'

which I've hopefully fixed by just changing the location of the
hard-coded path.  (I could also have fixed this by putting the new
location in secrets, but I figure this is harder for us to get
wrong. :-) )

Test Plan:
Fingers crossed.  Will deploy on toby and then type D1234 in the
infrastructure room and see what happens.  Will also do 'give 100
points to csilvers' and 'how many points does csilvers have' and hope
it's 100.

Reviewers: ben, james

Reviewed By: james

Differential Revision: https://phabricator.khanacademy.org/D14872