Add cache busting retry
Summary:
Per https://app.asana.com/0/27216215224639/34276303637220 the support sherpa does not pick up on newly created accounts until the next restart. This happens because it caches the list of users from Asana only once, at start-up.

This adds a retry in the case where the cache does not have the entry we expect to re-issue the query and update the cache.

And, man, this code is sooooo callbacky.

Test Plan: Kind of hard to test without adding a new HipChat user. I'll ping Karina to see how easy that is, but otherwise I'll just deploy and cross my fingers

Reviewers: csilvers

Reviewed By: csilvers

Subscribers: mroth

Differential Revision: https://phabricator.khanacademy.org/D18010