Fix brain issues to get plusplus script working hopefully
Summary:
It turns out that hubot loads the scripts in the `scripts/` folder in
alphabetical order. Because of this, the brain is loaded first
(`file-brain.coffee`) and then the plusplus script. However, that means
that the plusplus doesn't get to attach its listener to the brain before
the brain emits the `'loaded'` event. I do a manual reloading of the
brain when the plusplus script is loaded to get around this.

Test Plan:
- Deploy the bot
- Type @csilvers++ in a chat room
- The bot should output `@csilvers has <n> points`

Reviewers: csilvers

Subscribers: csilvers, james, ethan

Differential Revision: https://phabricator.khanacademy.org/D18737